VoicemailSettingsActivity: Fix some NPEs
If visual voicemail is supported, the PIN prefeerence is removed
when visual voice mail is disabled and re-added when enabled. This
causes NPEs in different parts of the code. Instead of adding the
needed checks not to crash, never remove the PIN preference if
visual voicemail is supported and simply enable/disable it.

BUGBASH-315

Change-Id: I6394c05cbd7adada1835e3cb89a535d6ffa9c2f4
Signed-off-by: DarkKnight6499 <yazad.madan@gmail.com>